requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiosqlite==0.19.0 